+++ packages/core/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the library will be documented in this file.
 
-## vX.X.X (Month DD, YYYY)
+## v0.6.2 (February 10, 2026)
 
 - Fix `PartialValues` type for array items